DANGEROUS: Workaround for variables (see full commit message)
This was an interesting challenge I tried to fix. It ultimately left me very confused as to how to fix it.

It seems the author had once declared some `final` variables somewhere before, and he was calling them somehow through the SpaceConfigHandler and Economy classes.

However, it seems like now those variables no longer exist or were moved somewhere else, and I could not find them.

So, my workaround was simply to remove the calls to the variables and instead just simply the returned value. I was extremely lenient about doing this, as I'm sure the author had a very distinct reasoning behind this, but I wasn't able to determine what that reasoning was.

Because of my wariness, I simply commented out the original lines and put my replacements in right above them. Should we ever figure this out one day, it should be an easy fix.

As a team, this is another discussion we should have, about the importance of not just "deleting" stuff without thinking, because as complex as this work is, it is crucial that we keep as much in tact as possible and make it our mission to only update for 1.8 and not add to it (yet).

That will hopefully come later.

/end commit