Debugging planet generation (see full msg)
The debug messages confirm that allowedShellIds and allowedCoreIds are
not empty by any means, the data is not transferred from there to the
Planetoid object for some reason... I'll have to investigate this some
more.

But at least it's loading from the files, and it seems to match the
materials just fine.