Refactor class to provide up-to-date information and make direct reference
Part of the main method was cleaned to remove the call to save the plugin prefix and version to a string. This class made a reference to those variables, and now, they call the plugin.yml for the information directly. The call was functional and complete in the Space class, but it did not seem necessary.